Move the itinerary page from [tripId] page
